Adds alignment property
In some cases, the dashed line would have an offset to the side, which
in return wouldn't be as pleseant to look at in comparison to the divider, since it is this widget that hopefully this package is trying to replace to.

So, by adding a property `alignment`, defaulted to `WrapAlignment.center`, one can achieve the desired alignment inside its application. If needed, this can be later changed by the developer to adjust for his/her scenario.

This may close the issue #25 - Feature: option to specify alignment.